pyproject.toml
[project]
name = "falsification-driven-rl"
version = "0.1.0"
description = "Implementation and comparison of Falsification-Driven Exploration in RL."
authors = [{ name = "Graham Alexander Noblit. Ph.D.", email = "grahamnoblit@pm.me" }]
requires-python = ">=3.10"

dependencies = [
    "gymnasium",
    "gymnasium-minigrid",
    "torch",
    "numpy",
    "tensorboard",
    "tqdm",
    "omegaconf",
    "hydra-core"
]

[project.scripts]
run-experiment = "falsify.main:main"

[tool.setuptools]
packages = ["falsify", "common"]